Mad Libs generator
This tells a tale about a person and his/her friend going for an adventure. You just enter words by following the instructions given, and it will give you a story based on my storyline. It is kind of random but it is funny although it might not make sense. Enjoy :)